Use webpack-merge to avoid duplicated configuration
