+++ package.json
@@ -130,6 +130,7 @@
     "watchify": "3.7.0",
     "webpack-bundle-analyzer": "^2.4.0",
     "webpack-delete-after-emit": "^1.0.5",
+    "webpack-merge": "^4.1.3",
     "webpack-stats-plugin": "0.1.4"
   },
   "optionalDependencies": {
