+++ webpack/common/modules.js
@@ -1,4 +1,4 @@
-const {resolve} = require('path');
+const { resolve } = require('path');
 
 module.exports = [
   'node_modules',
